refactor(keycloak): migrate KeycloakUserService and KeycloakGroupService to adapters
