+++ src/main/java/com/ebikes/iam/adapters/keycloak/KeycloakGroupAdapter.java
@@ -1,10 +1,11 @@
-package com.ebikes.iam.services.keycloak.groups;
+package com.ebikes.iam.adapters.keycloak;
 
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
+import jakarta.ws.rs.ProcessingException;
 import jakarta.ws.rs.core.Response;
 
 import org.keycloak.admin.client.Keycloak;
@@ -12,7 +13,7 @@ import org.keycloak.admin.client.resource.GroupsResource;
 import org.keycloak.admin.client.resource.RealmResource;
 import org.keycloak.admin.client.resource.UserResource;
 import org.keycloak.representations.idm.GroupRepresentation;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
 
 import com.ebikes.iam.configurations.properties.KeycloakProperties;
 import com.ebikes.iam.enums.ResponseCode;
@@ -24,63 +25,81 @@ import com.ebikes.iam.support.keycloak.KeycloakResponseUtilities;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
+@Component
 @RequiredArgsConstructor
-@Service
 @Slf4j
-public class KeycloakGroupService {
+public class KeycloakGroupAdapter {
 
-  private static final String KEYCLOAK_ENDPOINT = "keycloak://groups";
+  private static final String ENDPOINT = "keycloak://groups";
 
   private final Keycloak keycloakAdminClient;
   private final KeycloakProperties keycloakProperties;
 
   public void addUserToGroup(String groupPath, String userId) {
-    String groupId = getGroupIdOrThrow(groupPath);
-    getUserResource(userId).joinGroup(groupId);
-    log.info("User added to group - groupId={} groupPath={} userId={}", groupId, groupPath, userId);
+    try {
+      String groupId = getGroupIdOrThrow(groupPath);
+      getUserResource(userId).joinGroup(groupId);
+      log.info(
+          "User added to group - groupId={} groupPath={} userId={}", groupId, groupPath, userId);
+    } catch (ProcessingException e) {
+      throw new ExternalServiceException(
+          ENDPOINT, "Failed to add user to Keycloak group", ResponseCode.EXTERNAL_SERVICE_ERROR, e);
+    }
   }
 
   public String create(String organizationId, String organizationName) {
-    String groupPath = GroupPathUtilities.generate(organizationId);
-    Optional<GroupRepresentation> existing = findGroupByPath(groupPath);
-    if (existing.isPresent()) {
-      return existing.get().getId();
-    }
-
-    String groupName = groupPath.substring(1);
-    GroupRepresentation group = new GroupRepresentation();
-    group.setName(groupName);
-
-    Map<String, List<String>> attributes = new HashMap<>();
-    attributes.put("display_name", List.of(organizationName));
-    attributes.put("organization_type", List.of("customer"));
-    group.setAttributes(attributes);
-
-    try (Response response = getGroupsResource().add(group)) {
-      if (response.getStatus() != Response.Status.CREATED.getStatusCode()) {
-        throw new ExternalServiceException(
-            KEYCLOAK_ENDPOINT,
-            response.getStatusInfo().getReasonPhrase(),
-            ResponseCode.fromHttpStatus(response.getStatus()));
+    try {
+      String groupPath = GroupPathUtilities.generate(organizationId);
+      Optional<GroupRepresentation> existing = findGroupByPath(groupPath);
+      if (existing.isPresent()) {
+        return existing.get().getId();
       }
 
-      String groupId = KeycloakResponseUtilities.extractIdFromLocation(response);
-      log.info("Created group - groupId={} groupPath={}", groupId, groupPath);
-      return groupId;
+      String groupName = groupPath.substring(1);
+      GroupRepresentation group = new GroupRepresentation();
+      group.setName(groupName);
+
+      Map<String, List<String>> attributes = new HashMap<>();
+      attributes.put("display_name", List.of(organizationName));
+      attributes.put("organization_type", List.of("customer"));
+      group.setAttributes(attributes);
+
+      try (Response response = getGroupsResource().add(group)) {
+        String groupId = KeycloakResponseUtilities.extractIdFromLocation(response);
+        log.info("Created group - groupId={} groupPath={}", groupId, groupPath);
+        return groupId;
+      }
+    } catch (ProcessingException e) {
+      throw new ExternalServiceException(
+          ENDPOINT, "Failed to create Keycloak group", ResponseCode.EXTERNAL_SERVICE_ERROR, e);
     }
   }
 
   public String getGroupIdOrThrow(String groupPath) {
     return findGroupByPath(groupPath)
         .map(GroupRepresentation::getId)
-        .orElseThrow(() -> notFoundGroup(groupPath));
+        .orElseThrow(
+            () ->
+                new ResourceNotFoundException(
+                    ResponseCode.RESOURCE_NOT_FOUND, "Group not found: " + groupPath));
   }
 
   public void removeUserFromGroup(String groupPath, String userId) {
-    String groupId = getGroupIdOrThrow(groupPath);
-    getUserResource(userId).leaveGroup(groupId);
-    log.info(
-        "User removed from group - groupId={} groupPath={} userId={}", groupId, groupPath, userId);
+    try {
+      String groupId = getGroupIdOrThrow(groupPath);
+      getUserResource(userId).leaveGroup(groupId);
+      log.info(
+          "User removed from group - groupId={} groupPath={} userId={}",
+          groupId,
+          groupPath,
+          userId);
+    } catch (ProcessingException e) {
+      throw new ExternalServiceException(
+          ENDPOINT,
+          "Failed to remove user from Keycloak group",
+          ResponseCode.EXTERNAL_SERVICE_ERROR,
+          e);
+    }
   }
 
   private Optional<GroupRepresentation> findGroupByPath(String groupPath) {
@@ -97,11 +116,6 @@ public class KeycloakGroupService {
 
   private RealmResource realmResource() {
     return keycloakAdminClient.realm(keycloakProperties.getRealm());
-  }
-
-  private ResourceNotFoundException notFoundGroup(String groupPath) {
-    return new ResourceNotFoundException(
-        ResponseCode.RESOURCE_NOT_FOUND, "Group not found: " + groupPath);
   }
 
   private Optional<GroupRepresentation> searchInGroupHierarchy(
